modularize tracery modifiers; configurable logging
